move carousel components to sub folder for reuse
